LayoutTests/fast/files/revoke-blob-url-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: webkitURL
Test calling revokeObjectURL with no argument.
